feat(mapgen-studio-ui): B5 panels + splits (4; statusLabels + recipe-dag severed)
Move the 3 panels (ExplorePanel, GameConsole, RecipePanel) + the recipe-dag
directory split (PipelineStage owning the re-homed RecipeDagLoadStatus,
layout, domainPresentation, artifactPresentation, recipeDagFixture) + 4
stories + 5 relocated tests into @swooper/mapgen-studio-ui, and execute the
statusLabels split: the 3 formatters + runInGameRequiresProcessRestart +
the collapsed RunInGameRelation union land in panels/statusLabels.ts
(type-only over @civ7/studio-contract, E1-C); the app status modules keep
their constructors and alias the package union; dead PHASES re-exports die
app-side. useRecipeDagQuery back-imports the status union; StudioShell
repoints all panel + restart-predicate imports; ExplorePanel goes generic
over WaterStatsLayerRef so the app instantiates at its wider viz ref type.
buildRecipeDagLayout byte-parity is pinned by a fixture snapshot captured
from the app module at the parent tip. Export floor 90 -> 99 (exact);
45 stories cumulative package-side, AppHeader is the last app story.
