+++ apps/mapgen-studio/src/ui/components/index.ts
@@ -4,9 +4,7 @@
 // Re-exports all components for cleaner imports.
 // ============================================================================
 
-// Layout components
+// Layout components. (ExplorePanel/GameConsole/RecipePanel moved to
+// `@swooper/mapgen-studio-ui` — B5; AppHeader is the last app-side surface
+// component and moves with B6.)
 export { AppHeader, type AppHeaderProps } from "./AppHeader";
-export { ExplorePanel, type ExplorePanelProps } from "./ExplorePanel";
-export { GameConsole, type GameConsoleLiveRuntime, type GameConsoleProps } from "./GameConsole";
-// Panel components
-export { RecipePanel, type RecipePanelProps } from "./RecipePanel";
