+++ packages/mapgen-studio-ui/src/components/panels/recipe-dag/artifactPresentation.ts
@@ -1,4 +1,4 @@
-import { chooseRecipeDagDomainId, normalizeRecipeDagDomainId } from "./domainPresentation";
+import { chooseRecipeDagDomainId, normalizeRecipeDagDomainId } from "./domainPresentation.js";
 
 const ARTIFACT_TAG_PREFIX = /^artifact:/i;
